Transfer items to one special trough
Trough is being LF in AreasID.swill and will be leased to area
`harvest_area`. Means you need to have one area and `swill` with trough
inside of it. Before transfering to trouhg was looking for closest
trough and dumping it in `swill`, ifair i saw few conflicts with that.